+++ storage/dbshandler.go
@@ -1,10 +1,8 @@
 package storage
 
 import (
-	"fmt"
 	"gateserver/support"
 	"log"
-	"strings"
 )
 
 // handler for a DBS with generic value coming from a channel on a generic interface{}
@@ -44,9 +42,9 @@ func handlerDBS(id string, in chan interface{}, rst chan bool, a SampleData, tp 
 				//	fmt.Println("DEBUG DBS id:", id, "got data", d, "is current", support.StringEnding(id, "current", "_"))
 				//}
 				//TODO HERE being debugged
-				if strings.Contains(id, "hour") {
-					fmt.Println("DEBUG DBS id:", id, "got data", d, "is current", support.StringEnding(id, "current", "_"))
-				}
+				// if strings.Contains(id, "hour") {
+				// 	fmt.Println("DEBUG DBS id:", id, "got data", d, "is current", support.StringEnding(id, "current", "_"))
+				// }
 			case aa := <-rst:
 				// Reset via API might be dangerous, this is just a  placeholder
 				log.Println("storage.TimedIntDBS: handler id:", id, "got reset request", aa, " - does nothing still")